refactor: 타입 선언 변경
