+++ components/Card/Card.module.css
@@ -1,69 +1,150 @@
-.totalFrame {
-  display: flex;
-}
-.totalContainer {
-  display: flex;
-  flex-direction: column;
-  padding: 2rem;
-  width: 35.4rem;
-  height: auto;
-  border-right: 1px solid var(--gray20);
-}
-
-.headerContainer {
-  display: flex;
-  align-items: center;
-  margin-bottom: 2.5rem;
-  width: 100%;
-  height: fit-content;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.todoWrapper {
+.cardContainer {
   display: flex;
   gap: 1.2rem;
-  align-items: center;
-}
-
-.contentContainer {
-  display: flex;
-  gap: 1.6rem;
   flex-direction: column;
-}
-
-.buttonWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
+  padding: 2rem;
   border: 1px solid var(--gray30);
   width: 100%;
-  height: 4rem;
+  height: fit-content;
   background: var(--white);
   border-radius: 0.6rem;
 }
 
+.cardContentContainer {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+}
+.cardTitleWrapper {
+  margin-bottom: 1rem;
+  font-size: 1.6rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  color: var(--black1);
+}
+
+.allContentContainer {
+  display: flex;
+  flex-direction: column;
+}
+
+.cardTagContainer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.cardTagsWrapper {
+  display: flex;
+  gap: 0.6rem;
+  overflow: hidden;
+  white-space: nowrap;
+}
+
+.cardCountWrapper {
+  padding: 0.4rem 0.6rem;
+  width: fit-content;
+  height: 2rem;
+  background: var(--black2);
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: normal;
+  color: var(--gray20);
+  text-align: center;
+  border-radius: 0.8rem;
+}
+
+.cardInfoWrapper {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-top: 1.6rem;
+  flex-grow: 1;
+}
+
+.dateWrapper {
+  display: flex;
+  gap: 0.6rem;
+  align-items: center;
+}
+
+.date {
+  font-size: 1.2rem;
+  font-style: normal;
+  font-weight: 500;
+  line-height: normal;
+  color: var(--gray50);
+}
+
 @media screen and (max-width: 1200px) {
-  .totalFrame {
-    flex-direction: column;
+  .cardContainer {
+    flex-direction: row;
+    align-items: center;
   }
-  .totalContainer {
-    width: 58.4rem;
-    border-bottom: 1px solid var(--gray20);
-    border-right: none;
+  .imageContent {
+    width: 11rem;
+    height: auto;
+  }
+  .allContentContainer {
+    align-items: center;
+    flex-direction: row;
+  }
+  .cardTagsWrapper {
+    justify-content: start;
+    max-width: 20rem;
+  }
+  .cardTagCotainer {
+    margin-right: 1.6rem;
+  }
+  .cardCountWrapper {
+    display: none;
+  }
+  .cardInfoWrapper {
+    justify-content: space-between;
+    margin-top: 0;
   }
 }
 
 @media screen and (max-width: 767px) {
-  .totalFrame {
+  .cardContainer {
     flex-direction: column;
+    align-items: flex-start;
+    width: 100%;
   }
-  .totalContainer {
+  .cardContentContainer {
+    width: 100%;
+  }
+  .imageContent {
+    width: 26rem;
+    height: 15.2rem;
+  }
+  .allContentContainer {
     flex-direction: column;
-    width: 30.8rem;
-    margin-bottom: 1.2rem;
-    stroke-width: 1px;
-    stroke: var(--gray-gray_EEEEEE, #eee);
+    align-items: flex-start;
+  }
+  .cardTitleWrapper {
+    font-size: 1.4rem;
+  }
+  .cardTagCotainer {
+    margin: 0;
+    width: 100%;
+    gap: 0.8rem;
+  }
+  .dateIcon {
+    width: 1.4rem;
+    height: 1.4rem;
+  }
+  .date {
+    font-size: 1rem;
+  }
+  .cardInfoWrapper {
+    width: 100%;
+  }
+  .cardTagsWrapper {
+    max-width: 100%;
+  }
+  .cardCountWrapper {
+    display: flex;
   }
 }
