--- components/Card/card.type.ts
@@ -1,15 +1,0 @@
-export interface CardProps {
-  cardList: CardTaskProps[];
-}
-
-export interface CardTaskProps {
-  id?: number;
-  imageUrl: string | null;
-  title: string;
-  tags: string[];
-  dueDate: string;
-  assignee: {
-    nickname: string;
-    id: number;
-  };
-}
