+++ components/Chips/ChipColors/ChipColors.tsx
@@ -1,8 +1,17 @@
 import { useState } from "react";
 import style from "./ChipColors.module.css";
 import ChipColor from "../ChipColor/ChipColor";
-import { ChipColorsProps } from "@/components/Chips/chips.type";
-import { ChipColorProps } from "@/components/Chips/chips.type";
+
+interface ChipColorProps {
+  size: string;
+  color: "green" | "purple" | "orange" | "blue" | "pink";
+  onClick?: () => void;
+  isSelected: boolean;
+}
+
+interface ChipColorsProps {
+  size: "sm" | "lg";
+}
 
 const ChipColors = ({ size }: ChipColorsProps) => {
   const [selectedColor, setSelectedColor] = useState("green");
