+++ components/Table/TablePagination/TablePagination.tsx
@@ -1,4 +1,3 @@
-import { InviteBoard, Member, TableIndexType } from "@/components/Table/Table.type";
 import TableIndex from "@/components/Table/TableIndex/TableIndex";
 import TableList from "@/components/Table/TableList/TableList";
 import HideButton from "@/components/Table/TablePagination/HideButton";
@@ -9,10 +8,15 @@ import { clsx } from "clsx";
 import { useMemo, useState } from "react";
 import styles from "./TablePagination.module.css";
 import Image from "next/image";
+import { BasicUserType, InvitationData } from "@/types/api.type";
+
+type TableIndexType = {
+  [a: string]: "nickname" | "dashboard" | "inviter" | "email" | "deleteButton" | "acceptButton" | "cancelButton";
+};
 
 interface TableProps {
   title: string;
-  data: Member[] | InviteBoard[];
+  data: BasicUserType[] | InvitationData[];
   row?: number;
   tableIndex: TableIndexType;
   invite?: boolean;
